pyproject.toml
[project]
name = "protfrag"
version = "0.1.0"
description = "Protein fragment prediction from ProtT5 embeddings"
requires-python = ">=3.9"

dependencies = [
    "torch",
    "pytorch-lightning",
    "transformers",
    "pandas",
    "numpy",
    "scikit-learn",
    "biopython",
    "tqdm",
    "pyyaml",
    "matplotlib",
    "seaborn",
    "tensorboard"
]
